+++ perf-event/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added `Record::parse_sample_id` to allow accessing a `SampleId` when
+  `Builder::sample_id_all` is enabled.
 
 ## 0.7.4 - 2024-05-30
 ### Added